+++ cs4773-hw3/src/main/java/commands/shapes/Origin.java
@@ -1,27 +1,27 @@
 package commands.shapes;
 
 public class Origin {
-    double x;
-    double y;
+    int x;
+    int y;
 
-    public Origin(double x, double y) {
+    public Origin(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
 }
